+++ lib/src/sqlite_options.dart
@@ -1,3 +1,15 @@
+class WebSqliteOptions {
+  final String workerUri;
+  final String wasmUri;
+
+  const WebSqliteOptions.defaults()
+      : workerUri = 'db_worker.js',
+        wasmUri = 'sqlite3.wasm';
+
+  const WebSqliteOptions(
+      {this.wasmUri = 'sqlite3.wasm', this.workerUri = 'db_worker.js'});
+}
+
 class SqliteOptions {
   /// SQLite journal mode. Defaults to [SqliteJournalMode.wal].
   final SqliteJournalMode? journalMode;
@@ -11,6 +23,8 @@ class SqliteOptions {
   /// attempt to truncate the file afterwards.
   final int? journalSizeLimit;
 
+  final WebSqliteOptions webSqliteOptions;
+
   /// Timeout waiting for locks to be released by other connections.
   /// Defaults to 30 seconds.
   /// Set to null or [Duration.zero] to fail immediately when the database is locked.
@@ -20,12 +34,14 @@ class SqliteOptions {
       : journalMode = SqliteJournalMode.wal,
         journalSizeLimit = 6 * 1024 * 1024, // 1.5x the default checkpoint size
         synchronous = SqliteSynchronous.normal,
+        webSqliteOptions = const WebSqliteOptions.defaults(),
         lockTimeout = const Duration(seconds: 30);
 
   const SqliteOptions(
       {this.journalMode = SqliteJournalMode.wal,
       this.journalSizeLimit = 6 * 1024 * 1024,
       this.synchronous = SqliteSynchronous.normal,
+      this.webSqliteOptions = const WebSqliteOptions.defaults(),
       this.lockTimeout = const Duration(seconds: 30)});
 }